+++ src/test/java/racinggame/domain/CarTest.java
@@ -1,9 +1,10 @@
-package racinggame;
+package racinggame.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import racinggame.domain.value.Position;
 
 import java.util.stream.Stream;
