+++ src/test/java/racinggame/domain/value/MoveResultTest.java
@@ -1,9 +1,10 @@
-package racinggame.vo;
+package racinggame.domain.value;
 
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racinggame.Car;
+import racinggame.domain.Car;
+import racinggame.domain.value.MoveResult;
 
 import java.util.List;
