+++ src/test/java/racinggame/domain/value/PositionTest.java
@@ -1,8 +1,9 @@
-package racinggame;
+package racinggame.domain.value;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racinggame.domain.value.Position;
 
 import static org.assertj.core.api.Assertions.assertThat;
