+++ src/test/java/racinggame/service/RacingCarsTest.java
@@ -1,9 +1,11 @@
-package racinggame;
+package racinggame.service;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import racinggame.domain.Car;
+import racinggame.service.RacingCars;
 
 import java.util.ArrayList;
 import java.util.List;
